requirements.txt
alabaster==1.0.0
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-rest-swagger==2.2.0
djangorestframework==3.15.1
docutils==0.21.2
drf-spectacular=
idna==3.10
imagesize==1.4.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.1
openapi-codec==1.3.2
packaging==24.1
Pygments==2.18.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.3
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinx-autodoc-typehints==2.5.0
sphinx-rtd-theme==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
